refactor: 웹 dto 패키지 분리
