+++ src/main/java/nextstep/web/RoomEscapeController.java
@@ -1,6 +1,8 @@
 package nextstep.web;
 
 import nextstep.model.Reservation;
+import nextstep.web.dto.ReservationRequest;
+import nextstep.web.dto.ReservationResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -11,7 +13,6 @@ import java.net.URI;
 public class RoomEscapeController {
     private final RoomEscapeService roomEscapeService;
 
-    @Autowired
     public RoomEscapeController(RoomEscapeService roomEscapeService) {
         this.roomEscapeService = roomEscapeService;
     }
