+++ src/main/java/nextstep/web/dto/ReservationRequest.java
@@ -1,4 +1,4 @@
-package nextstep.web;
+package nextstep.web.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
